add Object class fix bug size_t

+++ object.h
@@ -1,15 +1,17 @@
 #ifndef OBJECT_H
 #define OBJECT_H
 
+#include <cstdlib>
+
 namespace DSLib
 {
 
 class Object
 {
 public:
-    void* operator new(unsigned int size) throw();
+    void* operator new(size_t size) throw();
     void operator delete(void* p);
-    void* operator new[](unsigned int size) throw();
+    void* operator new[](size_t size) throw();
     void operator delete[](void* p);
     virtual ~Object() = 0;
 };
